test recipe added && few mem functions implemented

+++ src/mem/ft_memcpy.c
@@ -1,11 +1,13 @@
-void	ft_memcpy(void *dest, void *src, unsigned long int n)	{
+#include <stddef.h>
+
+void	ft_memcpy(void *dest, void *src, size_t n)	{
 	char	*dest_ptr;
 	char	*src_ptr;
 
 	dest_ptr = dest;
 	src_ptr = src;
 	while (n)	{
-		*dest_ptr= *src_ptr;
+		*dest_ptr = *src_ptr;
 		dest_ptr++;
 		src_ptr++;
 		n--;
